requirements.txt
# pip==19.2.3
# bump2version==0.5.11
# wheel==0.33.6
# watchdog==0.9.0

# [development]
flake8==3.7.8
tox==3.14.0
coverage==4.5.4

pytest==6.2.4
black==21.7b0

# [production]
tweepy==4.5.0
python-dotenv==0.19.2
